Move inner classes to separate scripts
