+++ hsc/constants.py
@@ -1,0 +1,50 @@
+file_extensions = {
+	'ada': 'ada',
+	'bash': 'sh',
+	'c': 'c',
+	'clojure': 'clj',
+	'coffeescript': 'coffee',
+	'cpp': 'cpp',
+	'cpp14': 'cpp',
+	'csharp': 'cs',
+	'd': 'd',
+	'db2': 'sql',
+	'elixir': 'ex',
+	'erlang': 'erl',
+	'fortran': 'for',
+	'fsharp': 'fs',
+	'go': 'go',
+	'groovy': 'groovy',
+	'haskell': 'hs',
+	'java': 'java',
+	'java8': 'java',
+	'javascript': 'js',
+	'julia': 'jl',
+	'kotlin': 'kt',
+	'lolcode': 'lol',
+	'lua': 'lua',
+	'mysql': 'sql',
+	'objectivec': 'm',
+	'ocaml': 'ml',
+	'octave': 'oct',
+	'oracle': 'sql',
+	'pascal': 'pas',
+	'perl': 'pl',
+	'php': 'php',
+	'pypy': 'py',
+	'pypy3': 'py',
+	'python': 'py',
+	'python3': 'py',
+	'racket': 'rkt',
+	'r': 'r',
+	'ruby': 'rb',
+	'rust': 'rs',
+	'sbcl': 'lisp',
+	'scala': 'scala',
+	'swift': 'swift',
+	'smalltalk': 'st',
+	'tcl': 'tcl',
+	'tsql': 'sql',
+	'visualbasic': 'vbs',
+	'whitespace': 'hs',
+}
