+++ hsc/progress_bar.py
@@ -1,0 +1,4 @@
+from progress.bar import ChargingBar
+
+class CustomProgress(ChargingBar):
+	suffix = '%(percent)d%% [%(index)d/%(max)d]'
